timetracking/requirements.txt
django>=4.2,<6
django-environ>=0.11
django-bootstrap5>=23.3
psycopg2-binary>=2.9
